Update to adjust for lost photons
The central source should loose photons based on the ionizations as the radiation propagates out. This will look like an attenuation factor that is proportional to 1-n_ion/n_HII, where n_ion is the number density of ionizing photons and n_HII is the number density of ionized H atoms.  The bubbles now have roughly the right size -- please check! :)